+++ apstools/utils/__init__.py
@@ -10,6 +10,9 @@ from .device_info import listdevice
 
 from .email import EmailNotifications
 
+from .image_analysis import analyze_1D
+from .image_analysis import analyze_2D
+
 from .list_plans import listplans
 
 from .list_runs import getRunData